# Core: Producer Loop with Multi Parse Result Items 