--- a/application/apps/indexer/indexer_cli/src/interactive.rs
+++ b/application/apps/indexer/indexer_cli/src/interactive.rs
@@ -54,18 +54,28 @@ pub(crate) async fn handle_interactive_session(input: Option<PathBuf>) {
                                         println!("received shutdown through future channel");
                                         break;
                                     }
-                                    item = msg_stream.next() => {
+                                    items = msg_stream.next() => {
+                                        let items = match items {
+                                            Some(item) => item,
+                                            None => {
+                                                println!("no msg");
+                                                continue;
+                                            }
+                                        };
+                                        for item in items {
+
                                         match item {
-                                            Some((_, MessageStreamItem::Item(ParseYield::Message(msg)))) => {
+                                            (_, MessageStreamItem::Item(ParseYield::Message(msg))) => {
                                                 println!("msg: {msg}");
                                             }
-                                            Some((_, MessageStreamItem::Item(ParseYield::MessageAndAttachment((msg, attachment))))) => {
+                                            (_, MessageStreamItem::Item(ParseYield::MessageAndAttachment((msg, attachment)))) => {
                                                 println!("msg: {msg}, attachment: {attachment:?}");
                                             }
-                                            Some((_, MessageStreamItem::Item(ParseYield::Attachment(attachment)))) => {
+                                            (_, MessageStreamItem::Item(ParseYield::Attachment(attachment))) => {
                                                 println!("attachment: {attachment:?}");
                                             }
                                             _ => println!("no msg"),
+                                        }
                                         }
                                     }
                                 }