--- a/application/apps/indexer/parsers/src/lib.rs
+++ b/application/apps/indexer/parsers/src/lib.rs
@@ -35,16 +35,16 @@ impl<T> From<T> for ParseYield<T> {
 /// in chipmunk
 pub trait Parser<T> {
     /// take a slice of bytes and try to apply a parser. If the parse was
-    /// successfull, this will yield  the rest of the slice along with `Some(log_message)`
+    /// successful, this will yield the consumed bytes count along with `Some(log_message)`
     ///
     /// if the slice does not have enough bytes, an `Incomplete` error is returned.
     ///
     /// in case we could parse a message but the message was filtered out, `None` is returned
-    fn parse<'a>(
+    fn parse(
         &mut self,
-        input: &'a [u8],
+        input: &[u8],
         timestamp: Option<u64>,
-    ) -> Result<(&'a [u8], Option<ParseYield<T>>), Error>;
+    ) -> Result<(usize, Option<ParseYield<T>>), Error>;
 }
 
 #[derive(Debug, Clone, Serialize)]