--- a/application/apps/indexer/sources/src/producer/tests/mod.rs
+++ b/application/apps/indexer/sources/src/producer/tests/mod.rs
@@ -1,0 +1,5 @@
+mod mock_byte_source;
+mod mock_parser;
+mod single_parse;
+
+use super::*;