--- a/application/apps/indexer/sources/src/producer/tests/mod.rs
+++ b/application/apps/indexer/sources/src/producer/tests/mod.rs
@@ -1,5 +1,6 @@
 mod mock_byte_source;
 mod mock_parser;
+mod multi_parse;
 mod single_parse;
 
 use super::*;