This PR Changes the Producer Loop to support parsers returning multiple items at once on calling `parse()` method, which is essentially for implementing parsers in the plugins system.

Following has been implemented:
- [x] Parse returns the consumed count of bytes as `usize` instead of the reference to the remaining slice.
- [x] Parse returns an iterator of messages inside a result, to indicate that it should try parsing all messages in the given slice and return them or return an Error of no item has been parsed with the current call
- [x] Implementation of the current parsers hasn't changed much. They use `iter::once()` to indicate that they have one item only to get most compile time optimization from the compiler
- [x] Producer stream returns a boxed slice of all mapped items from `parse()` call, which gives the compiler the info that the consumer of them won't change it to get more optimizations.
- [x] Producer unit tests has been adjusted for `iter::once()` case without much change and extend with new tests with mock parsers returning multiple items at once
- [x] We have two benchmarks with mocks now instead of one (one for `iter::once()` and one for multiple values). Adjustments to the configuration of Build CLI tool are done too.
- [x] I've considered using `smallvec` but it had worse performance compared to boxed slice. 

Here are some benchmarks from DLT file (541 MB) running in Chipmunk app and in dlt benches of the current master and this branch for comparison + Benchmarks with mock parser returning single item in master and `iter::once()` in this Branch:

```Markdown

# Master: 

## App DLT:
File Read Took: 11.464 ms
File Read Took: 11.343 ms
File Read Took: 11.382 ms

## Bench DLT:
dlt_producer            time:   [4.0264 s 4.0323 s 4.0381 s]
dlt_producer            time:   [4.0504 s 4.0603 s 4.0698 s]

## Bench Mock (single):
mocks_producer/50000    time:   [5.0924 ms 5.1012 ms 5.1111 ms]
mocks_producer/50000    time:   [4.8922 ms 4.9009 ms 4.9105 ms]
mocks_producer/50000    time:   [5.4412 ms 5.4490 ms 5.4575 ms]

=========================================================

# PR:

## App DLT:
File Read Took: 11.430 ms
File Read Took: 11.309 ms
File Read Took: 11.261 ms

## Bench DLT:
dlt_producer            time:   [4.0641 s 4.0709 s 4.0781 s]
dlt_producer            time:   [4.0621 s 4.0713 s 4.0799 s]
dlt_producer            time:   [4.0772 s 4.0855 s 4.0936 s]

## Bench Mock (single):
mocks_producer/50000    time:   [4.3264 ms 4.3353 ms 4.3453 ms]
mocks_producer/50000    time:   [4.2533 ms 4.2647 ms 4.2769 ms]
mocks_producer/50000    time:   [4.2273 ms 4.2433 ms 4.2609 ms]

```